Config ui impovements added
