+++ client/src/auth/CandidateLayout.jsx
@@ -51,7 +51,7 @@ const CandidateLayout = () => {
     try {
       await dispatch(logoutCandidateAuth()).unwrap();
       showSuccessToast('Success', 'Logged out successfully');
-      navigate('/login');
+      navigate('/');
     } catch (error) {
       showErrorToast('Error', error.message || 'Logout failed');
       console.error('Logout failed:', error);
